[refact] Adapt code to have unique room
The algorithm provide the unicity of the edges but not the verteces.
To fix it we need to change the graph so we are sure that we can only
pass one time by room. For this we separate rooms in 2 and obligate to pass throw both

Before : start -> node -> end
Now    : start -> node -> node out -> end